Using adding edit to cabin creation

+++ src/services/apiCabins.js
@@ -22,37 +22,57 @@ export async function deleteCabin(id) {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-    '/',
-    ''
-  );
+export async function createEditCabin(newCabin, id = null) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  let imagePath;
+  let imageName;
 
-  // 1. Create cabin
-  const { data, error } = await supabase
-    .from('cabins')
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // If its editing a cabin and didnt change the image, just use the already setted url from image
+  if (hasImagePath) {
+    imageName = newCabin.image.split('/').at(-1);
+    imagePath = newCabin.image;
+  } else {
+    // else, if its a new cabin or changed the already existing image, use the file information and create a new image path
+    imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll('/', '');
+    imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  }
+
+  // 1. Create/edit cabin
+  let query = supabase.from('cabins');
+
+  // EDIT
+  if (id) {
+    console.log('editing');
+    query = query.update({ ...newCabin, image: imagePath }).eq('id', id);
+  }
+  // CREATE
+  else {
+    console.log('creating');
+    query = query.insert([{ ...newCabin, image: imagePath }]);
+  }
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
     throw new Error('Cabins could not be created');
   }
 
-  // 2. Upload images
-  const { error: storageError } = await supabase.storage
-    .from('cabin-images')
-    .upload(imageName, newCabin.image);
+  // 2. Upload images if an image was selected
+  if (!hasImagePath) {
+    const { error: storageError } = await supabase.storage
+      .from('cabin-images')
+      .upload(imageName, newCabin.image);
 
-  // 3. Delete the cabin if theres an error uploading the image
-  if (storageError) {
-    console.log(storageError);
-    await supabase.from('cabins').delete().eq('id', data.id);
-    throw new Error(
-      'Cabin image could not be uploaded and the cabin was not created'
-    );
+    // 3. Delete the cabin if theres an error uploading the image
+    if (storageError) {
+      console.log(storageError);
+      await supabase.from('cabins').delete().eq('id', data.id);
+      throw new Error(
+        'Cabin image could not be uploaded and the cabin was not created'
+      );
+    }
   }
 
   return data;
